["Initial" Commit]
This is not technically the initial commit because I worked on this on my own machine I did not upload. I finish this in the matter of days. One thing to note from this program is that it has a MAX byte size of 1,024 for a file copy. This bug is known and will be fixed at a later time.

-Tedster